+++ ui/src/dashboards/actions/index.js
@@ -18,6 +18,9 @@ import {
 import {
   notifyDashboardDeleted,
   notifyDashboardDeleteFailed,
+  notifyCellAdded,
+  notifyCellCloned,
+  notifyCellDeleted,
 } from 'shared/copy/notifications'
 
 import {
@@ -293,6 +296,7 @@ export const addDashboardCellAsync = dashboard => async dispatch => {
       getNewDashboardCell(dashboard)
     )
     dispatch(addDashboardCell(dashboard, data))
+    dispatch(notify(notifyCellAdded()))
   } catch (error) {
     console.error(error)
     dispatch(errorThrown(error))
@@ -306,6 +310,7 @@ export const cloneDashboardCellAsync = (dashboard, cell) => async dispatch => {
       getClonedDashboardCell(dashboard, cell)
     )
     dispatch(addDashboardCell(dashboard, data))
+    dispatch(notify(notifyCellCloned(cell.name)))
   } catch (error) {
     console.error(error)
     dispatch(errorThrown(error))
@@ -316,6 +321,7 @@ export const deleteDashboardCellAsync = (dashboard, cell) => async dispatch => {
   try {
     await deleteDashboardCellAJAX(cell)
     dispatch(deleteDashboardCell(dashboard, cell))
+    dispatch(notify(notifyCellDeleted(cell.name)))
   } catch (error) {
     console.error(error)
     dispatch(errorThrown(error))